Wallpaper is half black after rotating quickly
When rotate the device quickly in a transparent application activity,
sometime the wallpaper will be half black.

Root cause:
Sometime the transaction that change the size of the wallpaper surface
behind the drawFrame of wallpaper.

Redraw the frame of wallpaper one more time only when wallpaper
size changed can fix it.